error reporting, change API for cat command, and add description to caption

+++ main.py
@@ -18,32 +18,53 @@ def welcome_message(bot, message):
 
 @bot.on_message(filters.command("cat", prefixes=["/", ".", "!"]))
 def cat_generator(bot, message):
-    filename = "cat.jpg"
-    url = "https://cataas.com/cat"
-    photo = requests.get(url).content
-    open(filename, "wb").write(photo)
-    message.reply_photo(filename, quote=True)
-    os.remove(filename)
+    r = requests.get(
+        "https://api.thecatapi.com/v1/breeds/search?q=" + message.command[1:])
+    breed = r.json()["id"]
+    breedimageurl = requests.get(
+        "https://api.thecatapi.com/images/search?breed_id=" + breed)
+    if r.status_code != 200:
+        print("Error in Cat API")
+        bot.send_message(1833693304, "Error occured in Cat API!")
+        bot.reply_text(
+            "An error occured in the API, try again later or change your request.")
+    else:
+        photo = requests.get(breedimageurl.json()["url"]).content
+        filename = "cat.jpg"
+        open(filename, "wb").write(photo)
+        message.reply_photo(filename, quote=True,
+                            caption=breedimageurl.json()["description"])
+        os.remove(filename)
 
 
 @bot.on_message(filters.command("neko", prefixes=["/", ".", "!"]))
 def neko_generator(bot, message):
     r = requests.get("https://neko-love.xyz/api/v1/neko")
-    photo = requests.get(r.json()["url"]).content
-    filename = "neko.jpg"
-    open(filename, "wb").write(photo)
-    message.reply_photo(filename, quote=True)
-    os.remove(filename)
+    if r.status_code != 200:
+        print("Error in Neko API")
+        bot.send_message(1833693304, "Error occured in Neko API!")
+        bot.reply_text("An error occured in the API, try again later.")
+    else:
+        photo = requests.get(r.json()["url"]).content
+        filename = "neko.jpg"
+        open(filename, "wb").write(photo)
+        message.reply_photo(filename, quote=True)
+        os.remove(filename)
 
 
 @bot.on_message(filters.command("shiba", prefixes=["/", ".", "!"]))
 def shiba_generator(bot, message):
     r = requests.get("https://dog.ceo/api/breed/shiba/images/random")
-    photo = requests.get(r.json()["message"]).content
-    filename = "shiba.jpg"
-    open(filename, "wb").write(photo)
-    message.reply_photo(filename, quote=True)
-    os.remove(filename)
+    if r.status_code != 200:
+        print("Error in Shiba API")
+        bot.send_message(1833693304, "Error occured in Shiba API!")
+        bot.reply_text("An error occured in the API, try again later.")
+    else:
+        photo = requests.get(r.json()["message"]).content
+        filename = "shiba.jpg"
+        open(filename, "wb").write(photo)
+        message.reply_photo(filename, quote=True)
+        os.remove(filename)
 
 
 bot.run()
